+++ guillotina/contrib/oauth/discovery/__init__.py
@@ -1,0 +1,4 @@
+"""Discovery package — triggers registration of well-known handlers on import."""
+
+from guillotina.contrib.oauth.discovery import authorization_server as _as  # noqa
+from guillotina.contrib.oauth.discovery import protected_resource as _pr  # noqa
